+++ docs/conf.py
@@ -28,9 +28,9 @@ copyright = u'2013-2016, AlphaNodes GmbH'
 author = u'Alexander Meindl'
 
 # The short X.Y version.
-version = u'0.5.8'
+version = u'1.0.0'
 # The full version, including alpha/beta/rc tags.
-release = u'0.5.8'
+release = u'1.0.0'
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.